refactor(tracing): extract the Tracer SPI into a new kpipe-tracing module
kpipe-tracing-otel previously did `requires transitive io.github.eschizoid.kpipe.producer`
solely to name the Tracer SPI, dragging the whole producer module onto every tracing user's
module path. The SPI (Tracer + its nested SpanScope) now lives in a tiny standalone module:

- New module lib/kpipe-tracing (JPMS io.github.eschizoid.kpipe.tracing), requiring only
  kafka.clients (transitive — the signatures expose ConsumerRecord/Headers). Published,
  BOM-pinned, and released via the existing graph-derived wiring (no manual list edits).
- kpipe-producer, kpipe-consumer, and kpipe-api each add
  `requires transitive io.github.eschizoid.kpipe.tracing` (they all name Tracer in public
  API) with matching Gradle `api(project(":lib:kpipe-tracing"))` edges.
- kpipe-tracing-otel now requires only the SPI module; its dependency on kpipe-producer is
  gone entirely (verified: no kpipe-producer on its runtime classpath).

Migration (delete-and-migrate, no deprecation cycle):
- Package moved: io.github.eschizoid.kpipe.producer.tracing -> io.github.eschizoid.kpipe.tracing.
  Fix is a one-line import change; all in-repo callers migrated in this commit.
- kpipe-tracing-otel users no longer pull kpipe-producer transitively. If your code used
  producer types through that edge (unlikely), add an explicit kpipe-producer dependency.
