+++ lib/kpipe-consumer/build.gradle.kts
@@ -8,6 +8,7 @@ description = "KPipe Consumer - Functional Kafka consumer with virtual threads"
 dependencies {
   api(project(":lib:kpipe-core"))
   api(project(":lib:kpipe-producer"))
+  api(project(":lib:kpipe-tracing"))
 
   implementation(libs.kafkaClients)
