+++ lib/kpipe-consumer/src/main/java/io/github/eschizoid/kpipe/consumer/RecordProcessor.java
@@ -3,9 +3,9 @@ package io.github.eschizoid.kpipe.consumer;
 import io.github.eschizoid.kpipe.metrics.ConsumerMetricKeys;
 import io.github.eschizoid.kpipe.metrics.ConsumerMetrics;
 import io.github.eschizoid.kpipe.producer.KPipeProducer;
-import io.github.eschizoid.kpipe.producer.tracing.Tracer;
 import io.github.eschizoid.kpipe.registry.MessagePipeline;
 import io.github.eschizoid.kpipe.registry.Result;
+import io.github.eschizoid.kpipe.tracing.Tracer;
 import java.lang.System.Logger;
 import java.lang.System.Logger.Level;
 import java.nio.channels.ClosedByInterruptException;
