+++ lib/kpipe-producer/src/main/java/io/github/eschizoid/kpipe/producer/package-info.java
@@ -7,9 +7,8 @@
 /// - [io.github.eschizoid.kpipe.producer.config] — immutable producer-side configuration.
 /// - [io.github.eschizoid.kpipe.producer.sink]   — the Kafka-backed
 /// [io.github.eschizoid.kpipe.sink.MessageSink] implementation.
-/// - [io.github.eschizoid.kpipe.producer.tracing] — pluggable
-/// [io.github.eschizoid.kpipe.producer.tracing.Tracer] SPI for cross-Kafka-boundary tracing.
 ///
-/// OpenTelemetry-backed tracing is provided by `kpipe-tracing-otel`; this module ships only
-/// the SPI and a no-op default.
+/// Cross-Kafka-boundary tracing is pluggable via the
+/// [io.github.eschizoid.kpipe.tracing.Tracer] SPI (the `kpipe-tracing` module, required
+/// transitively); the OpenTelemetry-backed implementation is the opt-in `kpipe-tracing-otel`.
 package io.github.eschizoid.kpipe.producer;
