+++ lib/kpipe-producer/src/main/java/module-info.java
@@ -2,10 +2,11 @@
 module io.github.eschizoid.kpipe.producer {
   requires transitive io.github.eschizoid.kpipe.core;
   requires transitive io.github.eschizoid.kpipe.metrics;
+  // transitive: KPipeProducer / KafkaMessageSink expose Tracer in their public API.
+  requires transitive io.github.eschizoid.kpipe.tracing;
   requires transitive kafka.clients;
 
   exports io.github.eschizoid.kpipe.producer;
   exports io.github.eschizoid.kpipe.producer.config;
   exports io.github.eschizoid.kpipe.producer.sink;
-  exports io.github.eschizoid.kpipe.producer.tracing;
 }
