+++ settings.gradle.kts
@@ -5,6 +5,7 @@ include("lib:kpipe-bom")
 include("lib:kpipe-core")
 include("lib:kpipe-metrics")
 include("lib:kpipe-metrics-otel")
+include("lib:kpipe-tracing")
 include("lib:kpipe-tracing-otel")
 include("lib:kpipe-schema-registry-confluent")
 include("lib:kpipe-consumer")
